rename: rename loading components for enhanced seonjo experience
