+++ frontend/src/core/showLoading.js
@@ -1,6 +1,6 @@
 import { initializeRouter, changeUrl, parsePath } from "./router.js";
 import { getCookie } from "./jwt.js";
-import { createLoadingElement, addLoadingStyles } from "./loading.js";
+import { createLoadingElement, addLoadingStyles } from "./loadingComponents.js";
 
 const WEBSOCKET_URL = 'wss://localhost:443/ws/online/';
